downloader/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
anyio==4.0.0
    # via httpx
beautifulsoup4==4.12.2
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
exceptiongroup==1.1.3
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
loguru==0.7.2
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
tqdm==4.66.1
    # via -r requirements.in
